+++ middleware/comicDefined.js
@@ -1,7 +1,6 @@
 export default defineNuxtRouteMiddleware(() => {
     const comicStore = useComicStore();
-    console.log(comicStore.comic);
-    console.log(comicStore.comic.getPage(0));
+
     if (
         !(
             comicStore.comic.getPage(0) &&
