+++ pages/export.vue
@@ -180,7 +180,6 @@
     function saveDraft() {
         let comicJson = comicStore.comic.toJSON();
         comicStore.saveDraft(comicJson);
-        console.log('comic is saved');
     }
 
     onMounted(() => {
